Completion of task number 6, link below.
https://benja.orangescrum.com/dashboard#/details/aa72ddd6b0682f328a2a7c85d1ffdadf

A new FileInteraction.js class has been created. Within this class is a
new method named getRootDirectory that returns the root directory for
a given user in the form of an array of document objects. These document
objects are not something we've defined, or need to, it is just how
mongoDB structures the documents within the collection.